Fix UI behaviour on captcha start and solve
On some screen resolutions, `this.elements.codeEditor.focus()` fails
to focus the captcha (for whatever reason), which results in it
loading off-screen. To fix this, I've added a scroll animation that
manually scrolls to the captcha UI.

To match the behaviour of standard captchas, the captcha now hides
itself when solved.